requirements.txt
Flask
requests
squareup
